refactor: depend on pybluetti instead of embedding the API client
Step 3 of the pybluetti extraction (see chpego/pybluetti): the
integration now depends on the published pybluetti>=0.1.0 package for
all BLUETTI cloud communication (HTTP + websocket push updates)
instead of embedding its own copy under api/, model/product.py, and
application_exception.py - all deleted here.

- manifest.json requirements: pybluetti>=0.1.0 replaces pydantic,
  stomper, and websocket-client directly (now pybluetti's own
  transitive dependencies).
- APPLICATION_PROFILE (the region/server-URL config singleton) moves
  from the now-deleted api/bluetti.py to profile/application_profile.py,
  where the class it instantiates already lived.
- Every ProductClient/StompClient construction site (__init__.py,
  oauth.py, options_flow.py) now passes the resolved gateway/websocket
  URL and an on_auth_expired callback explicitly, matching pybluetti's
  hass-decoupled constructor signature - preserving today's exact
  behavior (every 805 response still fires EVENT_TOKEN_EXPIRED on the
  hass event bus, regardless of which call site hit it).
- StompClient.connect()/disconnect() are now coroutines (pybluetti's
  async websocket transport, no more dedicated threads) - both call
  sites updated to await them.
- Method/StringEnum dropped from const.py (moved to pybluetti.const in
  the extraction's first step; nothing here used them outside the
  deleted api/ package).
- quality_scale.yaml: async-dependency flipped from todo to done - the
  integration's only non-HA dependency is now pybluetti, which is
  fully async end to end.

No user-visible behavior change. 100% line coverage maintained (tests
for the deleted modules removed - they're superseded by pybluetti's
own already-ported suite; tests exercising these constructors updated
for the new signatures/async mocks).
