--- custom_components/bluetti/api/__init__.py
@@ -1,7 +1,0 @@
-from .bluetti import Bluetti
-from .product_client import ProductClient
-
-__all__ = [
-    "Bluetti",
-    "ProductClient"
-]
