+++ custom_components/bluetti/config_flow.py
@@ -3,11 +3,11 @@
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import callback
 
-from .api.bluetti import APPLICATION_PROFILE
 from .application_credentials import async_ensure_default_credential
 from .const import DOMAIN
 from .oauth import OAuth2FlowHandler
 from .options_flow import BluettiOptionsFlowHandler
+from .profile.application_profile import APPLICATION_PROFILE
 
 
 class BluettiConfigFlow(OAuth2FlowHandler, domain=DOMAIN):
