+++ custom_components/bluetti/const.py
@@ -1,5 +1,4 @@
 """Constants for the BLUETTI integration."""
-from enum import Enum
 
 DOMAIN: str = "bluetti"
 INTEGRATION_NAME: str = "BLUETTI"
@@ -14,17 +13,3 @@ NOTIFY_ID_TOKEN_EXPIRED: str = "notifyTokenExpire"  # noqa: S105 - notification 
 # standard duplicate-prevention mechanism instead of matching on the entry
 # title.
 ACCOUNT_UNIQUE_ID: str = "account"
-
-class StringEnum(str, Enum):
-    """String Enum define."""
-
-    def __str__(self) -> str:
-        return self.value
-
-
-class Method(StringEnum):
-    """HTTP Methods define."""
-
-    GET = "GET"
-    POST = "POST"
-    DELETE = "DELETE"
