+++ custom_components/bluetti/manifest.json
@@ -13,6 +13,6 @@
   "iot_class": "cloud_push",
   "issue_tracker": "https://github.com/bluetti-official/bluetti-home-assistant/issues",
   "quality_scale": "gold",
-  "requirements": ["pydantic>=2.0", "stomper>=0.4", "websocket-client>=1.6"],
+  "requirements": ["pybluetti>=0.1.0"],
   "version": "1.1.0"
 }
