--- custom_components/bluetti/model/product.py
@@ -1,13 +1,0 @@
-
-from pydantic import BaseModel
-
-
-class UserProduct(BaseModel):
-    """"""
-    sn: str
-    stateList: list
-    online: str
-    model: str | None = None
-    name: str | None = None
-    isBindByCurUser: str | None = None
-
