+++ custom_components/bluetti/oauth.py
@@ -13,8 +13,8 @@ from homeassistant.helpers import config_entry_oauth2_flow
 from homeassistant.helpers import issue_registry as ir
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.event import async_track_time_interval
+from pybluetti import ProductClient
 
-from .api.product_client import ProductClient
 from .const import (
     ACCOUNT_UNIQUE_ID,
     DOMAIN,
@@ -22,6 +22,7 @@ from .const import (
     INTEGRATION_NAME,
     NOTIFY_ID_TOKEN_EXPIRED,
 )
+from .profile.application_profile import APPLICATION_PROFILE
 
 __LOGGER__ = logging.getLogger(__name__)
 
@@ -110,7 +111,12 @@ class OAuth2FlowHandler(config_entry_oauth2_flow.AbstractOAuth2FlowHandler, doma
 
         httpSession = async_get_clientsession(self.hass)
         access_token = self._oauth_data["token"]["access_token"]
-        product_client = ProductClient(httpSession, access_token,self.hass)
+        product_client = ProductClient(
+            httpSession,
+            APPLICATION_PROFILE.config["server"]["gateway"],
+            access_token,
+            on_auth_expired=lambda: self.hass.bus.fire(EVENT_TOKEN_EXPIRED),
+        )
         try:
             products = await product_client.get_user_products()
         except Exception as err:
