+++ custom_components/bluetti/quality_scale.yaml
@@ -76,6 +76,6 @@ rules:
   stale-devices: done
 
   # Platinum
-  async-dependency: todo
+  async-dependency: done
   inject-websession: done
   strict-typing: todo
